# PEX: fluid event log