@yarikk 2021-12-09
```
% go test
…
PASS
ok  	github.com/anacrolix/torrent	2.273s
```
```
% go version
go version go1.17.4 darwin/arm64
```