@yarikk 2021-12-09
Oh, I see, fails under `-count 2`. I'll fix the test.