@yarikk 2021-12-09
I see, you've got `-count 2` in the CI run. I fixed the test to keep the test data good for second iteration.