@anacrolix 2021-12-10
https://github.com/anacrolix/torrent/issues/691#issuecomment-990612549