@anacrolix 2021-12-11
Memory usage seems negligible. Let's merge.